requirements.txt
fastapi==0.128.0
uvicorn== 0.40.0
pillow==12.1.0
python-multipart==0.0.22